CLEM clean up individual attributes ready for linking to properties
