+++ Models/CLEM/Resources/IndividualAttributeList.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Models.CLEM.Interfaces;
 
 namespace Models.CLEM.Resources
 {
